requirements.txt
flask
tensorflow
numpy
pillow
opencv-python
matplotlib
ultralytics
gunicorn